refactor(mcp): 拆分 MCP 服务边界
